+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Prop `className` to `AddressForm` component.
+
 ## [3.12.14] - 2020-09-23
 
 ### Fixed